actualizacion modulo compras final

+++ app/Controllers/Producto.php
@@ -3,6 +3,7 @@ namespace App\Controllers;
 use App\Controllers\BaseController;
 use App\Models\Productos;
 use App\Models\Categorias;
+use App\Models\DetalleCompras;
 
 class Producto extends BaseController{
     public function index(){
@@ -82,10 +83,21 @@ class Producto extends BaseController{
     }
     
     public function eliminar($id = null) {
-        $producto = new Productos();
-        $datos = $producto->where('id', $id)->first();
-        $producto->where('id', $id)->delete($id);
-        return $this->response->redirect(site_url('productos'));
+        $detalles = new DetalleCompras();
+        //validar si hay categorias con productos relacionados
+        $detalle_compra = $detalles->where('id_producto', $id)->findAll();
+        if(!empty($detalle_compra)){
+            $session = session();
+            $session->setFlashData('mensaje','No se puede eliminar el producto ya que tiene transacciones relacionadas.');
+            return $this->response->redirect(site_url('productos'));
+        }else{
+            $producto = new Productos();
+            $datos = $producto->where('id', $id)->first();
+            $producto->where('id', $id)->delete($id);
+            return $this->response->redirect(site_url('productos'));
+        }
+
+        
 
     }
